Documents/Projects/trading_bot_template/requirements.txt
greenlet==3.1.1
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
pandas
numpy
requests
beautifulsoup4
tensorflow
scikit-learn
backtrader